chore: MemoDetailViewController 파일에 있던 열거형 새로운 파일로 위치 변경
